blocked off admin pages from non-admin users

+++ app/Http/Controllers/AdminController.php
@@ -7,6 +7,7 @@ use App\Award;
 use App\User;
 use App\Course;
 use App\Category;
+use Auth;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
@@ -37,44 +38,74 @@ ON nomination.id=course.nomination_id Where nomination_id  in (SELECT id from no
 
 
   public function awardReport(){
-
+    if (Auth::user()->admin===1) {
       $awards = Award::all();
       $nominations = Nomination::all();
       $unique_Years = DB::select('SELECT EXTRACT(YEAR FROM created_at) AS uniqueYears FROM nomination group by uniqueYears');
       $countNoms = DB::select('SELECT  count(id) AS countID ,award_id from nomination group by award_id');
       return view('admin.awardReport')->with('nominations', $nominations)->with('awards', $awards)->with('unique_Years' ,$unique_Years)->with('countNoms',$countNoms);
+      }
+        else {
+          return view('pages.noAccess');
+        }
     }
 
     public function getReportByYear(Request $request){
-      $awards = Award::all();
-      $nominations = Nomination::all();
-      $unique_Years = DB::select('SELECT EXTRACT(YEAR FROM created_at) AS uniqueYears FROM nomination group by uniqueYears  ');
-      $countNoms = DB::select('SELECT EXTRACT(YEAR FROM created_at) AS uniqueYears, count(id) AS countID ,award_id from nomination group by award_id HAVING uniqueYears IN ');
-      return view('admin.awardReport')->with('nominations', $nominations)->with('awards', $awards)->with('unique_Years' ,$unique_Years)->with('countNoms',$countNoms);
+      if (Auth::user()->admin===1) {
+        $awards = Award::all();
+        $nominations = Nomination::all();
+        $unique_Years = DB::select('SELECT EXTRACT(YEAR FROM created_at) AS uniqueYears FROM nomination group by uniqueYears  ');
+        $countNoms = DB::select('SELECT EXTRACT(YEAR FROM created_at) AS uniqueYears, count(id) AS countID ,award_id from nomination group by award_id HAVING uniqueYears IN ');
+        return view('admin.awardReport')->with('nominations', $nominations)->with('awards', $awards)->with('unique_Years' ,$unique_Years)->with('countNoms',$countNoms);
+      }
+      else {
+          return view('pages.noAccess');
+        }
     }
 
     public function award() {
+      if (Auth::user()->admin===1) {
         $awards = Award::all();
         $categories = Category::all();
         return view('admin.award')->with('awards', $awards)->with('categories', $categories);
+      }
+      else {
+          return view('pages.noAccess');
+      }
     }
 
     public function prof() {
+      if (Auth::user()->admin===1) {
         $profs = User::all();
         return view('admin.prof')->with('profs', $profs);
       }
+      else {
+          return view('pages.noAccess');
+      }
+    }
 
     public function search() {
+      if (Auth::user()->admin===1) {
         return view('admin.search');
+      }
+      else {
+          return view('pages.noAccess');
+      }
     }
 
 	public function nominations() {
-        $awards = Award::all();
-        $courses = course::all();
-        $nominations = Nomination::all();
-        $unique_Years = DB::select('SELECT EXTRACT(YEAR FROM created_at) AS uniqueYears FROM nomination group by uniqueYears');
 
-        return view('admin.nominations')->with('nominations', $nominations)->with('awards',$awards)->with('courses', $courses)->with('unique_Years',$unique_Years);
+        if (Auth::user()->admin===1) {
+          $awards = Award::all();
+          $courses = course::all();
+          $nominations = Nomination::all();
+          $unique_Years = DB::select('SELECT EXTRACT(YEAR FROM created_at) AS uniqueYears FROM nomination group by uniqueYears');
+
+          return view('admin.nominations')->with('nominations', $nominations)->with('awards',$awards)->with('courses', $courses)->with('unique_Years',$unique_Years);
+        }
+        else {
+          return view('pages.noAccess');
+        }
     }
 
 
@@ -83,20 +114,31 @@ ON nomination.id=course.nomination_id Where nomination_id  in (SELECT id from no
 
   /*  This is the NomineeInfo Tab*/
   public function nomineeInfo() {
+    if (Auth::user()->admin===1) {
         $nominees = Nominee::all();
         return view('admin.nomineeInfo')->with('nominees', $nominees);
     }
+    else {
+          return view('pages.noAccess');
+    }
+  }
 
   public function editEmail($studentNumber) {
 
+      if (Auth::user()->admin===1) {
         // get the nominee info including the email
         $nominee = Nominee::find($studentNumber);
 
         // show edit form and pass on Email
         return view('admin.editEmail')->with('nominee', $nominee);
-    }
+      }
+      else {
+          return view('pages.noAccess');
+      }
+  }
 
   public function updateEmail(Request $request, $studentNumber) {
+    if (Auth::user()->admin===1) {
         $this->validate($request, [
             'email'=>'required',
             ]);
@@ -106,8 +148,14 @@ ON nomination.id=course.nomination_id Where nomination_id  in (SELECT id from no
 
         $nominees = nominee::all();
         return view('admin.nomineeInfo')->with('nominees', $nominees);
-    }
+        }
+      else {
+          return view('pages.noAccess');
+      }
+  }
+
   public function storeEmail(Request $request) {
+    if (Auth::user()->admin===1) {
         $this->validate($request, [
             'email'=>'required',
             ]);
@@ -117,16 +165,26 @@ ON nomination.id=course.nomination_id Where nomination_id  in (SELECT id from no
 
         $nominees = nominee::all();
         return view('admin.nomineeInfo')->with('nominees', $nominees);
+        }
+      else {
+          return view('pages.noAccess');
+      }
     }
 /*--------------------------------------------------------------------------*/
 /*BRANDON STUFF*/
 
   public function emailTemplate(){
+    if (Auth::user()->admin===1) {
         $nominees = Nominee::all();
         return view('admin.emailTemplate');
+        }
+      else {
+          return view('pages.noAccess');
+      }
     }
 
   public function sendEmail(){
+    if (Auth::user()->admin===1) {
         $nominees = Nominee::all(); 
         //loop through each nominee
         foreach ($nominees as $nominee) {
@@ -142,11 +200,16 @@ ON nomination.id=course.nomination_id Where nomination_id  in (SELECT id from no
 
         }
         return view('admin.emailSent')->with('nominees', $nominees);
+        }
+      else {
+          return view('pages.noAccess');
+      }
     }
 
 /*----------------------*/
 
     public function storeAward(Request $request) {
+      if (Auth::user()->admin===1) {
         $this->validate($request, [
             // 'award'=>'required',
             'name'=>'required',
@@ -164,6 +227,10 @@ ON nomination.id=course.nomination_id Where nomination_id  in (SELECT id from no
         $awards = Award::all();
         $categories = Category::all();
         return view('admin.award')->with('awards', $awards)->with('categories', $categories);
+        }
+      else {
+          return view('pages.noAccess');
+      }
 
     }
 
@@ -174,12 +241,17 @@ ON nomination.id=course.nomination_id Where nomination_id  in (SELECT id from no
      * @return \Illuminate\Http\Response
      */
     public function editAward($id) {
+      if (Auth::user()->admin===1) {
         // get the award
         $award = Award::find($id);
         $categories = Category::all();
 
         // show edit form and pass on award
         return view('admin.editAward')->with('award', $award)->with('categories', $categories);
+        }
+      else {
+          return view('pages.noAccess');
+      }
     }
 
     /**
@@ -190,6 +262,7 @@ ON nomination.id=course.nomination_id Where nomination_id  in (SELECT id from no
      * @return \Illuminate\Http\Response
      */
     public function updateAward(Request $request, $id) {
+      if (Auth::user()->admin===1) {
         $this->validate($request, [
             // 'award'=>'required',
             'name'=>'required',
@@ -205,6 +278,10 @@ ON nomination.id=course.nomination_id Where nomination_id  in (SELECT id from no
         $awards = Award::all();
         $categories = Category::all();
         return view('admin.award')->with('awards', $awards)->with('categories', $categories);
+        }
+      else {
+          return view('pages.noAccess');
+      }
     }
 
     /**
@@ -214,17 +291,23 @@ ON nomination.id=course.nomination_id Where nomination_id  in (SELECT id from no
      * @return \Illuminate\Http\Response
      */
     public function destroyAward($id) {
+      if (Auth::user()->admin===1) {
         $award = Award::find($id)->delete();
 
         $awards = Award::all();
         $categories = Category::all();
 
         return view('admin.award')->with('awards', $awards)->with('categories', $categories);
+        }
+      else {
+          return view('pages.noAccess');
+      }
     }
 
 
 
     public function storeProf(Request $request) {
+      if (Auth::user()->admin===1) {
         $this->validate($request, [
             'firstName'=>'required',
             'lastName'=>'required',
@@ -236,17 +319,27 @@ ON nomination.id=course.nomination_id Where nomination_id  in (SELECT id from no
 
         $profs = User::all();
         return view('admin.prof')->with('profs', $profs);
+        }
+      else {
+          return view('pages.noAccess');
+      }
     }
 
     public function editProf($id) {
+      if (Auth::user()->admin===1) {
         // get the prof
         $prof = User::find($id);
 
         // show edit form and pass on prof
         return view('admin.editProf')->with('prof', $prof);
+        }
+      else {
+          return view('pages.noAccess');
+      }
     }
 
     public function updateProf(Request $request, $id) {
+      if (Auth::user()->admin===1) {
         $this->validate($request, [
             'firstName'=>'required',
             'lastName'=>'required',
@@ -260,6 +353,10 @@ ON nomination.id=course.nomination_id Where nomination_id  in (SELECT id from no
         // return redirect()->route('award.report');
         $profs = User::all();
         return view('admin.prof')->with('profs', $profs);
+        }
+      else {
+          return view('pages.noAccess');
+      }
     }
 
     /**
@@ -269,26 +366,42 @@ ON nomination.id=course.nomination_id Where nomination_id  in (SELECT id from no
      * @return \Illuminate\Http\Response
      */
     public function destroyProf($id) {
+      if (Auth::user()->admin===1) {
         $prof = User::find($id)->delete();
 
         $profs = User::all();
         return view('admin.prof')->with('profs', $profs);
+        }
+      else {
+          return view('pages.noAccess');
+      }
     }
 
     public function Categories() {
+      if (Auth::user()->admin===1) {
         $categories = Category::all();
         return view('admin.categories')->with('categories', $categories);
+        }
+      else {
+          return view('pages.noAccess');
+      }
     }
 
     public function editCategory($id) {
+      if (Auth::user()->admin===1) {
         // get the category
         $category = Category::find($id);
 
         // show edit form and pass on category
         return view('admin.editCategory')->with('category', $category);
+        }
+      else {
+          return view('pages.noAccess');
+      }
     }
 
     public function updateCategory(Request $request, $id) {
+      if (Auth::user()->admin===1) {
 
       $this->validate($request, [
             'name'=>'required',
@@ -301,9 +414,14 @@ ON nomination.id=course.nomination_id Where nomination_id  in (SELECT id from no
         // return redirect()->route('award.report');
         $categories = Category::all();
         return view('admin.categories')->with('categories', $categories);
+        }
+      else {
+          return view('pages.noAccess');
+      }
     }
 
     public function storeCategory(Request $request) {
+      if (Auth::user()->admin===1) {
       $this->validate($request, [
             'name'=>'required',
             ]);
@@ -313,11 +431,20 @@ ON nomination.id=course.nomination_id Where nomination_id  in (SELECT id from no
 
         $categories = Category::all();
         return view('admin.categories')->with('categories', $categories);
+        }
+      else {
+          return view('pages.noAccess');
+      }
     }
 
     public function destroyCategory() {
+      if (Auth::user()->admin===1) {
         $category = Category::find($id)->delete();
         $categories = Category::all();
         return view('admin.categories')->with('categories', $categories);
+        }
+      else {
+          return view('pages.noAccess');
+      }
     }
 }
